Create Player.java
Player object is created at start of program. This class will represent the players character. Player will have an inventory and keep track of the room they are in. Player will have a weight limit of items it can pick up. Player may eventually have other stats or objects which it will need to record  and keep track of.