Update Game.java
A lot of changes here. I am in process of refactorring this class and attempting to make it more cohesive. originally extended this class with Actions.java but couldnt seem to update the current room the player was in correctly. I think making extensions from the command class for each individual movement and command will provide better modularity and control.

Began working to implement a Stack<Room> to keep track of the history of rooms used to implement a back feature. not sure if this is where it will stay.

Added goRoom back to this class hoping it would fix the issue of player realizing it is in the current room not the original room. It lived in Actions.java for awhile. Both seems like the wrong way to do it.

a lot of code here is commented out because im still shuffling around the logic with how to implement it.