update on Message.java and comment on Items.java
added a Player player variable and  get/set methods for it incase I need to reference to player from here.
After all the code shifting today, I am not sure this class is going to stay as it is. I would still like a place to put all the long text to be called into game when needed. Working on that...