Update Player.java
HashMap is seeming more appropriate for the Inventory on the player as opposed to the HashSet only because of the extra value which a String can be used to hold it's name and I am not worrying about non static variables (Im not sure if related but I was having that trouble and now Im not.) key is String name, value is Items object

Began the Stack<Room> roomHistory /back functionality in here too. I think this might be there better place to put it as oppsed to Game class. will ch,