Huge Refactor. Edited Command.java. Added many more classes
Due to trouble with the old code in keeping the current room readable in every class it was established that the code wasnt very cohesive. So, in the effort to be both cohesive and modular we've begun a huge refactor.

Command class has become abstrack and holds all the actions in it's extended classes.  Action.java will become obsolete and phased out.
Once the commands are all working it should be easier to add another action by just created a new extended command class.
Parser.java also had to be refactored in order to no longer look in CommandWord enum but now hold the second word to be called and used by commands. (at least thats the idea behind logic right now.)