Large Refactor!
A lot of this was updating naming conventions and modularizing each class.
There is now better cohesion where almost all classes have only one task. Commands are now handled as calling extensions of the abstract Command class.
Created some classes as place holders for when they are implemented.
There are now classes which will create the objects en masse in the game named create_* .
The NPC class which will hold characters is starting to be implemented.
The history of room movedments recorded in roomStack is starting to be implemented.
cmd_Help, cmd_Quit, cmd_Items,  and cmd_Move have now all been implemented.
There is now a place for locked doors to be implemented inside the Room class.

The player can now move from room to room.
Next implementation will work on better Item controls and a working history stack of rooms. There will also start to be characters addded and the create_* classes will be fleshed out.