Update cmd_Look.java
added boolean to test if item can be described.
added ability to describe different objects in the room through a serious of boolean tests. Since each object will be unique, the test runs throgugh Items in room then Items in inventory of player and then checks to see if it is an NPC. When one of those tests positive then their description is listed and search stops.