updated the readme
updated the README.TXT file to be a legitimate quick summary and also for a view/ as a reference for where the game is currently at and outlined the next few steps I am going to take for the game.

added a rough ASCII floor plan to explain how the General layout of the floors are.